Fix deleteposition bug

+++ src/main/java/seedu/address/logic/commands/DeletePositionCommand.java
@@ -15,6 +15,8 @@ public class DeletePositionCommand extends Command {
     public static final String COMMAND_WORD = "deleteposition";
     public static final String MESSAGE_SUCCESS = "Position %s has been deleted successfully!";
     public static final String MESSAGE_NOT_FOUND = "Position doesn't exist.";
+    public static final String MESSAGE_POSITION_ASSIGNED =
+            "Cannot delete position %s as it is currently assigned to one or more players.";
     public static final String MESSAGE_MISSING_FLAG = "Missing 'ps/' flag for deleteposition command";
     public static final String MESSAGE_INVALID_FORMAT =
             "Invalid command format. Please ensure correct form at: deleteposition ps/<position_name>";
@@ -39,13 +41,21 @@ public class DeletePositionCommand extends Command {
         if (!Position.isValidPositionName(name)) {
             throw new CommandException(Position.MESSAGE_CONSTRAINTS);
         }
+
+        final Position toDelete;
         try {
-            Position toDelete = model.getPositionByName(name);
-            model.deletePosition(toDelete);
-            return new CommandResult(String.format(MESSAGE_SUCCESS, name));
+            toDelete = model.getPositionByName(name);
         } catch (RuntimeException e) {
             throw new CommandException(MESSAGE_NOT_FOUND);
         }
+
+        // Check if the position is assigned to any player
+        if (model.isPositionAssigned(toDelete)) {
+            throw new CommandException(String.format(MESSAGE_POSITION_ASSIGNED, name));
+        }
+
+        model.deletePosition(toDelete);
+        return new CommandResult(String.format(MESSAGE_SUCCESS, name));
     }
 
     @Override
